requirements.txt
black==22.6.0
click==8.1.3
mypy-extensions==0.4.3
mysql-connector-python==8.0.30
pathspec==0.9.0
platformdirs==2.5.2
protobuf==3.20.1
tabulate==0.8.10
tomli==2.0.1
